+++ src/util/fst_impl/bytes_ref_fst_enum.rs
@@ -159,7 +159,7 @@ where
 
     fn grow(&mut self) -> Result<()> {
         self.base.take_do_return(|base| {
-            ArrayUtil::grow_with_len(&mut *self.current.bytes.borrow_mut(), base.upto as i32 + 1);
+            ArrayUtil::grow_with_len(&mut *self.current.bytes.borrow_mut(), base.upto + 1);
             Ok(())
         })
     }
