+++ src/util/long_bit_set.rs
@@ -40,7 +40,7 @@ impl LongBitSet {
             let num_words = Self::bits2words(num_bits)?;
             let length = bits.bits.len();
             if num_words as usize >= length {
-                ArrayUtil::grow_with_len(&mut bits.bits, num_words + 1);
+                ArrayUtil::grow_with_len(&mut bits.bits, (num_words + 1) as usize);
             }
             debug_assert!(bits.bits.len() <= i32::MAX as usize);
             bits.num_bits = (bits.bits.len() as i64) << 6;
